Add doc-validator workflow
`doc-validator` codifies best practices for technical documentation at
Grafana.
Linting errors are defined in
https://github.com/grafana/technical-documentation/blob/main/tools/cmd/doc-validator/errata.hcl.

The errors enumerated in the errata file are under active development
and are intended to be helpful in explaining reasoning and fixes for
linting issues.

Not all errors meet this goal yet, so please report any errors that
are confusing, seem to lack motivation, or are hard to resolve.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>